+++ src/vehicle/vehiclecommand/StepperST.cpp
@@ -8,5 +8,5 @@ StepperST::StepperST(CommandRegistry ident, StepperFront &stepper, double multi)
         : VehicleCommand(ident), stepper(stepper), multi(multi) {}
 
 void StepperST::execute(const int16_t *params) {
-    stepper.setTarget((params[0]) * multi);
+    stepper.setTarget((static_cast<int16_t>((params[0]) * multi)));
 }
